+++ Backend/controllers/user-controller.js
@@ -17,7 +17,7 @@ const getUserDetail = async (req, res) => {
     const userDetail = await User.findOne({
       where: { id },
     });
-    if (userDetail != null) {
+    if (userDetail) {
       return res.status(200).send(userDetail);
     } else {
       return res.status(404).send("not found");
@@ -31,6 +31,16 @@ const getUserDetail = async (req, res) => {
 const createUser = async (req, res) => {
   try {
     const { Username, Password, Phone, Role, Email, Address } = req.body;
+    //ktra Username
+    const checkUsername = await User.findOne({ where: { Username } });
+    if (checkUsername) {
+      return res.status(400).send("đã tồn tại username");
+    }
+    //ktra email
+    const checkEmail = await User.findOne({ where: { Email } });
+    if (checkEmail) {
+      return res.status(400).send("đã tồn tại email");
+    }
     //tạo chuỗi ngẫu nhiên
     const salt = bcryptjs.genSaltSync(10);
     //mã hóa passs + salt
@@ -54,13 +64,20 @@ const createUser = async (req, res) => {
 const deleteUser = async (req, res) => {
   try {
     const { id } = req.params;
-    await User.destroy({
-      where: {
-        id,
-      },
+    const userDetail = await User.findOne({
+      where: { id },
     });
-    const list = await User.findAll();
-    res.status(200).send(list);
+    if (userDetail) {
+      await User.destroy({
+        where: {
+          id,
+        },
+      });
+      const list = await User.findAll();
+      res.status(200).send(list);
+    } else {
+      return res.status(404).send("not found");
+    }
   } catch (error) {
     res.status(500).send(error);
   }
@@ -71,16 +88,28 @@ const updateUser = async (req, res) => {
   try {
     const { id } = req.params;
     const { Phone, Role, Email, Address } = req.body;
-    await User.update(
-      { Phone, Role, Email, Address },
-      {
-        where: {
-          id,
-        },
-      }
-    );
-    const list = await User.findAll();
-    res.status(200).send(list);
+    //ktra email
+    const checkEmail = await User.findOne({ where: { Email } });
+    if (checkEmail) {
+      return res.status(400).send("đã tồn tại email");
+    }
+    const userDetail = await User.findOne({
+      where: { id },
+    });
+    if (userDetail) {
+      await User.update(
+        { Phone, Role, Email, Address },
+        {
+          where: {
+            id,
+          },
+        }
+      );
+      const list = await User.findAll();
+      res.status(200).send(list);
+    } else {
+      return res.status(404).send("not found");
+    }
   } catch (error) {
     res.status(500).send(error);
   }
@@ -90,12 +119,14 @@ const updateUser = async (req, res) => {
 const changePass = async (req, res) => {
   try {
     const { Username, Password, NewPassword } = req.body;
+    //kiểm tra user
     const UserLogin = await User.findOne({
       where: {
         Username,
       },
     });
     if (UserLogin != null) {
+      //ktra pass cũ
       const isAuth = bcryptjs.compareSync(Password, UserLogin.Password);
       if (isAuth) {
         //tạo chuỗi ngẫu nhiên
